Separate app for tagging schema
This is just cleaner that way. This is really no a concern of the blog app.
The blog app makes used of some classes that are defined in the nre
tagging.schema but that is it.

Special tagging for the blog pages (i.e. the BlogPagTag class) remain
in the blog app.