Remove blog query
Before, I though I needed a special query to return the connection
objects for a given tag, which then give me access to the blog pages
of the given tag.

It turns out, this is not needed. Once there is a graphql field on the
blog page that defines the connection back to the `BlogPageTag` objects,
then somehow graphene_django picks this up and knows how to display the
connection to the `BlogPageTag` on the `taggit.models.Tag`.

I had noticed before, that the `Tag` model has a field
`blog_blogpagetag_items` which links back to from the `Tag` to the
`BlogPageTag`.

So now you can find the blog pages that are tagged with a given tag like
so:
```graphql
{
  tag(name: "first") {
    id
    name
    blogBlogpagetagItems {
      contentObject {
        title
      }
    }
  }
}
```

I think some renaming would not hurt just to make things clearer.