Add nginx service config with restart policy
Before, I was experiencing issues because nginx where nginx was not available after restart.
I think this has something to do with the mountin of the shared /vagrant directory
in which the site configuration lives. I think it is still good to make sure that the
service tries to restart a few times when it fails on the first attempt.