Ignore local git repo when copying to image
Before I was seeing a git error complaining about missing host verfication key.
I guess that is the key I have locally set up to communicate with GitHub.

This can be cirumvernted by not copying over the repo.

It does raise the question how this could be done with a private repo though.