Serve media from /media not /cms/media
For some reason the django development serving of the media files was breaking when
media and static files have the same URL prefix... Not sure why that is...
But this is the quick work around.

So now /cms and /media are seved from the backend. Everything else should be caught
by the frontend.