Simplify provisioning
Now most action that needs to be done on the server is combined into
one simple script that needs to be run by the admin.

The location where the archive is extracted to does not matter any more.
The provisioning is flexible enough to set the right permissions anyways.

The automatically create user `dockrunner` is now automatically in
restricted mode and the necessary commands are installed during
user creation.