Switch ssh-add command back to file version
The other one did lead to weird error messages because the identity file was not found
but the remote commands still worked. I am guessing this has something to do
with the ssh-agent, but I am not sure. This file way of doing it is just way more transparent
for me.