Add extended document query type to GraphQL schema
The extended type adds a `src` field to the document. This
is a fully qualified URL under which the respective document
can be accessed on the backend.

Before, it was necessay to construct the document URL on the
client side. Since that is error prone and depended on the
backend configuration, this query/field was added.

Now the frontend only needs to use the extended query and will
always receive a usable URL.