Ensure enviroment variable are available to cron
Before, I was trying to get the Netlify build hook to fire when a scheduled page
is published. What I got was that the page would be set to published in Wagtail,
but not build was triggered in Netlify.

When publishing pages manually, the build hook would work as desired and Netlify would
build the new page.

The publishing of the scheduled pages is triggered by a cron job. The logs of that cron
job showed that when the `manage.py pulblish_scheduled_pages` is run, that the
environment variable for the build hook is missing.

[This answer](https://unix.stackexchange.com/a/27291/414381) points out that environment
variables that you rely on in you shell are not availble in cron jobs, because they are run
by the system with a minimal environment.

I already have a `.env` file on the host. The defined variables in there are made available
to the shell through docker. Because these variables are not available to the cron job
I am mounting the `.env` file into the container and sourcing it in the cron job script.
The sourcing might not be necessary, because the `manage.py` script is loading the `.env`
file with python-dotenv anyhow. But this is a reproducible way that can be used in other commands
which can not user python-dotenv and load the variables themselves.